training/requirements.txt
boto3
torch
transformers
diffusers
Pillow
torchvision
